+++ pytorch_lightning/callbacks/progress/rich_progress.py
@@ -18,7 +18,6 @@ from typing import Any, Optional, Union
 
 from pytorch_lightning.callbacks.progress.base import ProgressBarBase
 from pytorch_lightning.utilities import _RICH_AVAILABLE
-from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
 Task, Style = None, None
 if _RICH_AVAILABLE:
@@ -196,7 +195,7 @@ class RichProgressBar(ProgressBarBase):
 
     Args:
         refresh_rate_per_second: the number of updates per second. If refresh_rate is 0, progress bar is disabled.
-        display_every_n_epochs: Set to a non-negative integer to display progress bar every n epochs.
+        leave: Displays progress bar per epoch. Default: False
         theme: Contains styles used to stylize the progress bar.
 
     Raises:
@@ -207,7 +206,7 @@ class RichProgressBar(ProgressBarBase):
     def __init__(
         self,
         refresh_rate_per_second: int = 10,
-        display_every_n_epochs: Optional[int] = None,
+        leave: bool = False,
         theme: RichProgressBarTheme = RichProgressBarTheme(),
     ) -> None:
         if not _RICH_AVAILABLE:
@@ -216,11 +215,7 @@ class RichProgressBar(ProgressBarBase):
             )
         super().__init__()
         self._refresh_rate_per_second: int = refresh_rate_per_second
-        if display_every_n_epochs and (not isinstance(display_every_n_epochs, int) or (display_every_n_epochs < 0)):
-            raise MisconfigurationException(
-                f"`display_every_n_epochs` should be an int >= 0, got {display_every_n_epochs}."
-            )
-        self._display_every_n_epochs: Optional[int] = display_every_n_epochs
+        self._leave: bool = leave
         self._enabled: bool = True
         self.progress: Optional[Progress] = None
         self.val_sanity_progress_bar_id: Optional[int] = None
@@ -331,11 +326,7 @@ class RichProgressBar(ProgressBarBase):
         total_batches = total_train_batches + total_val_batches
 
         train_description = self._get_train_description(trainer.current_epoch)
-        if (
-            self.main_progress_bar_id is not None
-            and self._display_every_n_epochs
-            and (not trainer.current_epoch % self._display_every_n_epochs)
-        ):
+        if self.main_progress_bar_id is not None and self._leave:
             self._stop_progress()
             self._init_progress(trainer, pl_module)
         if self.main_progress_bar_id is None: